+++ tests/integration_tests/test_http_connection.py
@@ -339,8 +339,11 @@ class TestHttpConnection(TestCase):
             res = connection.session_info()
             self.assertFalse(res.is_error(), res)
             self.assertIsNotNone(res.query)
+            self.assertIsNotNone(res.result['session_id'])
+            deleted = res.result
+            deleted.pop('session_id')
             self.assertEqual({'db': 'test', 'http_origin': None, 'ip': '127.0.0.1', 'ns': 'test',
-                              'access': None, 'session_id': None}, res.result)
+                              'access': None}, deleted)
 
     def test_remove_table_with_record(self):
         surreal = Surreal(URL, credentials=('root', 'root'), use_http=True)