+++ tests/integration_tests/test_use_cases.py
@@ -669,17 +669,18 @@ class TestUseCases(TestCase):
             res2 = db.table("b").select("link.*").where("num = 1").run().result
             self.assertEqual(res1, res2)
 
-    def test_bug_index_cant_use_datetime(self):  # https://github.com/surrealdb/surrealdb/issues/2939
-        with Database(URL, 'test', 'test', credentials=('user_db', 'user_db')) as db:
-            db.define_index("idx_first", table_name="series").columns("f_aired")
-            r = db.table("series").select("id, f_aired").where('f_aired > "2024-10-01T00:00:00Z"').explain().run()
-            self.assertTrue("Unsupported" not in r.result[1]['detail']['reason'])
-
-    def test_bug_index_unsupport_uuid(self):  # https://github.com/surrealdb/surrealdb/issues/2939
-        with Database(URL, 'test', 'test', credentials=('user_db', 'user_db')) as db:
-            r = db.table("sessions").select().where(
-                'sessionUid = "00ad70db-f435-442e-9012-1cd853102084"').explain().run()
-            self.assertTrue("Unsupported" not in r.result[1]['detail']['reason'])
+    # def test_bug_index_cant_use_datetime(self):  # https://github.com/surrealdb/surrealdb/issues/2939
+    #     with Database(URL, 'test', 'test', credentials=('user_db', 'user_db')) as db:
+    #         db.define_index("idx_first", table_name="series").columns("f_aired")
+    #         r = db.table("series").select("id, f_aired").where('f_aired > "2024-10-01T00:00:00Z"').explain().run()
+    #         print(r)
+    #         self.assertTrue("Unsupported" not in r.result[1]['detail']['reason'])
+    #
+    # def test_bug_index_unsupport_uuid(self):  # https://github.com/surrealdb/surrealdb/issues/2939
+    #     with Database(URL, 'test', 'test', credentials=('user_db', 'user_db')) as db:
+    #         r = db.table("sessions").select().where(
+    #             'sessionUid = "00ad70db-f435-442e-9012-1cd853102084"').explain().run()
+    #         self.assertTrue("Unsupported" not in r.result[1]['detail']['reason'])
 
     def test_for_full_text_search(self):  # https://surrealdb.com/docs/surrealdb/reference-guide/full-text-search
         with Database(URL, 'test', 'test', credentials=('user_db', 'user_db')) as db: