+++ ReleaseNotes.md
@@ -19,6 +19,7 @@
 - add DEFINE ACCESS JWT and DEFINE ACCESS RECORD statements
 - add REMOVE ACCESS statement
 - add ALTER TABLE statement
+- add ENFORCED statement for DEFINE TABLE
 - deprecate DEFINE TOKEN-REMOVE TOKEN and DEFINE SCOPE-REMOVE SCOPE
 - let and unset now raises CompatibilityError for http
 - add upsert method to Connection